Copy website files to planktos/files
Essentially all this code was for line sw.js:20 which does an early
exit if the file requested is for the webseed. The early exit tells
the browser to handle the http request and not the service worker
which has significant performance improvements.

This allows the service worker to early exit on requests that
start with /planktos/files/, but still allow the other planktos
files like /planktos/planktos.min.js.