Move fetch() to lib (#38)
* Move fetch() to lib

track #36

* Rename scope to root

Scope represented the scope of the service worker, the directory
that the sw resided in. It is more intuitive to call it the root
because that directory is also the planktos root directory,
and most of the time the website's root.

* Copy website files to planktos/files

Essentially all this code was for line sw.js:20 which does an early
exit if the file requested is for the webseed. The early exit tells
the browser to handle the http request and not the service worker
which has significant performance improvements.

This allows the service worker to early exit on requests that
start with /planktos/files/, but still allow the other planktos
files like /planktos/planktos.min.js.

* Add more debug logging code to the lib

* Add comments and small fixes