Add web stream support
Web stream is used by default in fetch() if the browser supports
it. It can be disabled by setting opts.stream = false.

The way files are prioritized changed because ranges of a file
can be requested so instead of storing the file's name/hash,
the prioritized byte range is stored instead.

closes #29