Store snapshots as an ordered list
By representing the planktos instance as an ordered list of
snapshots a lot of edge cases are avoided. Most revolve around
the notion of the 'latest' snapshot since this implies that
there is always a latest and is distinguished from other snapshots
in some way. Representing as a list implies zero or more elements
which more closely resembles planktos since it can have zero or
more snapshots. Ordering provides a way to to determine the latest
snapshot while avoiding the edge cases that previously existed.
Ordering also provides a way to tell which snapshots are newer,
which allows for more versatillity in the lib e.g. a user could
delete all but the 2 newest snapshots.