Add ability to remove snapshots
Deleting a snapshot removes it from a planktos instances local
snapshot list, stops the seeding of the snapshot, and removes it from
the snapshot idb-kv-store. This does not interfer with other instances
who also have the same snapshot except in the case that the snapshot
was being downloaded which it will not continue.