Change api to export class and expose multiple snapshots (#51)
* Export a class in index.js

This allows Planktos to be instantiated with the `new` keyword
instead of global instance of planktos. All globals are now
gone from the codebase

* Add namespace option to Planktos constructor

This allows different instances to live in their own namespace
which insulates them from instances in a different namespace.
This is important because planktos instances share underlying
resources which can cause one instances to interfere with a
seperate instance

* Expose getAllSnapshots() in planktos instance

Planktos inheritly supports multiple snapshots due to keeping
old snapshots around, and now the old and previously hidden
snapshots can be accessed via the public api.

* Store snapshots as an ordered list

By representing the planktos instance as an ordered list of
snapshots a lot of edge cases are avoided. Most revolve around
the notion of the 'latest' snapshot since this implies that
there is always a latest and is distinguished from other snapshots
in some way. Representing as a list implies zero or more elements
which more closely resembles planktos since it can have zero or
more snapshots. Ordering provides a way to to determine the latest
snapshot while avoiding the edge cases that previously existed.
Ordering also provides a way to tell which snapshots are newer,
which allows for more versatillity in the lib e.g. a user could
delete all but the 2 newest snapshots.

* Add ability to remove snapshots

Deleting a snapshot removes it from a planktos instances local
snapshot list, stops the seeding of the snapshot, and removes it from
the snapshot idb-kv-store. This does not interfer with other instances
who also have the same snapshot except in the case that the snapshot
was being downloaded which it will not continue.

* Remove getSnapshot() in favor of getAllSnapshots()

The removal is part of enforcing the ordered list of snapshots
notion

* Fixed relative string urls being passed into fetch()

* Add api documentation

* Fix docs url

* Remove docs as submodule

* Stop using promises with transactions

Firefox and Chrome implement promises different which prevents
promises from being used with IndexedDB transactions because
when a promise reaction is handled the transaction is already
finished on Firefox. Chrome hanles the reaction in a way that
allows transactions to be used before they finished.