Remove info on what hash is used
There is confusion surrounding how planktos is effected by the
shattered sha1. Since the source torrent (contains sha1 hashes)
is created by the trusted website owner the shattered issue does
not apply to planktos' case. Shattered sha1 is only viable if the
attacher can generate both colliding files wich is not the case
for planktos since the website owner is trusted and creates the
torrent.